+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@haensl/iso-log",
-  "version": "3.1.1",
+  "version": "3.1.2",
   "description": "Isomorphic logger with Bunyan and Sentry support.",
   "type": "module",
   "exports": {
@@ -62,8 +62,7 @@
   "homepage": "https://github.com/haensl/iso-log#readme",
   "files": [
     "index.mjs",
-    "index.cjs",
-    "iso-log.js"
+    "index.cjs"
   ],
   "dependencies": {
     "@haensl/log": "^1.3.11"
